#40 feat: 유저 페르소나 enum 생성, refactor: enum -> entity 패키지로 이동
