+++ src/main/java/org/sopt/solply_server/domain/auth/service/OAuthService.java
@@ -1,6 +1,6 @@
 package org.sopt.solply_server.domain.auth.service;
 
-import org.sopt.solply_server.domain.auth.constant.SocialPlatform;
+import org.sopt.solply_server.domain.auth.entity.SocialPlatform;
 import org.sopt.solply_server.domain.user.entity.User;
 
 public interface OAuthService {
