--- src/main/java/org/sopt/solply_server/domain/tag/constant/TagName.java
@@ -1,4 +1,0 @@
-package org.sopt.solply_server.domain.tag.constant;
-
-public enum TagName {
-}
