+++ src/main/java/org/sopt/solply_server/domain/tag/entity/TagType.java
@@ -1,0 +1,4 @@
+package org.sopt.solply_server.domain.tag.entity;
+
+public enum TagType {
+}
