+++ src/main/java/org/sopt/solply_server/domain/user/service/SocialUserService.java
@@ -2,7 +2,7 @@ package org.sopt.solply_server.domain.user.service;
 
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.sopt.solply_server.domain.auth.constant.SocialPlatform;
+import org.sopt.solply_server.domain.auth.entity.SocialPlatform;
 import org.sopt.solply_server.domain.user.entity.SocialUserInfo;
 import org.sopt.solply_server.domain.user.entity.User;
 import org.sopt.solply_server.domain.user.repository.SocialUserInfoRepository;
